Move CLI_ARGS to main to avoid it requiring args as lib
